订单管理

+++ src/main/java/com/zzuli/housekeepingserver/service/impl/CategoryServiceImpl.java
@@ -6,6 +6,7 @@ import com.zzuli.housekeepingserver.bean.extend.CategoryExtend;
 import com.zzuli.housekeepingserver.dao.CategoryMapper;
 import com.zzuli.housekeepingserver.dao.extend.CategoryExtendMapper;
 import com.zzuli.housekeepingserver.service.CategoryService;
+import com.zzuli.housekeepingserver.utils.CustomerException;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
@@ -16,7 +17,7 @@ import java.util.List;
  * Modify Information:
  * Author        Date          Description
  * ============ =========== ============================
- * liang         2021/6/2
+ * liang         2021/6/2       栏目业务实现类
  */
 
 @Service
@@ -34,7 +35,7 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public void saveOrUpdate(Category category) {
+    public void saveOrUpdate(Category category) throws CustomerException {
         if (category.getId() != null) {
             categoryMapper.updateByPrimaryKey(category);
         } else {
@@ -43,7 +44,14 @@ public class CategoryServiceImpl implements CategoryService {
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteById(Long id) throws CustomerException {
+        // 先判断该id对应的数据存在不？
+        Category category = categoryMapper.selectByPrimaryKey(id);
+        if (category == null) {
+            // 当不存在，报错！删除
+            throw new CustomerException("删除失败,要删除的数据不存在");
+        }
+        // 当存在，删除
         categoryMapper.deleteByPrimaryKey(id);
     }
