+++ src/main/java/com/zzuli/housekeepingserver/service/impl/UserServiceImpl.java
@@ -6,6 +6,7 @@ import com.zzuli.housekeepingserver.bean.extend.UserExtend;
 import com.zzuli.housekeepingserver.dao.UserMapper;
 import com.zzuli.housekeepingserver.dao.extend.UserExtendMapper;
 import com.zzuli.housekeepingserver.service.UserService;
+import com.zzuli.housekeepingserver.utils.CustomerException;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
@@ -16,7 +17,7 @@ import java.util.List;
  * Modify Information:
  * Author        Date          Description
  * ============ =========== ============================
- * liang         2021/6/3
+ * liang         2021/6/3       用户业务实现类
  */
 
 @Service
@@ -34,7 +35,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void saveOrUpdate(User user) {
+    public void saveOrUpdate(User user) throws CustomerException {
         if (user.getId() != null) {
             userMapper.updateByPrimaryKey(user);
         } else {
@@ -43,7 +44,14 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteById(Long id) throws CustomerException {
+        // 先判断该id对应的数据存在不？
+        User user = userMapper.selectByPrimaryKey(id);
+        if (user == null) {
+            // 当不存在，报错！删除
+            throw new CustomerException("删除失败,要删除的数据不存在");
+        }
+        // 当存在，删除
         userMapper.deleteByPrimaryKey(id);
     }
